upd cli
