+++ internal/api/api.go
@@ -18,11 +18,13 @@ func NewAPI(
 	downloader Downloader,
 	monitor Monitor,
 	router *gin.Engine,
+	host string,
 ) *API {
 	return &API{
 		downloader: downloader,
 		monitor:    monitor,
 		router:     router,
+		host:       host,
 	}
 }
 
@@ -31,10 +33,14 @@ type API struct {
 	monitor    Monitor
 
 	router *gin.Engine
+	host   string
 }
 
-func (api *API) Run(ctx context.Context) {
+func (api *API) Run(ctx context.Context) error {
 	apiHandlers := api.router.Group("/api")
+
 	apiHandlers.POST("/create_task", api.createTask)
 	apiHandlers.GET("/task_status", api.taskStatus)
+
+	return api.router.Run(api.host)
 }
